src/backend/scrapers/facebook_real_scraper.py
#!/usr/bin/env python3
"""
Facebook Real Data Scraper - Extract actual Facebook posts
NO MOCK DATA - Only real posts from Facebook
"""

import json
import time
import re
from typing import Dict, List, Optional
from datetime import datetime
import os
from dotenv import load_dotenv
from src.backend.core.logging_config import logger
from src.backend.validators.real_data_validator import enforce_real_data

load_dotenv()


class FacebookRealScraper:
    """Scraper for real Facebook data - public posts only"""

    def __init__(self):
        self.session_data = {"posts_collected": [], "locations_found": set(), "groups_searched": []}

    def search_facebook_groups(self, group_urls: List[str]) -> List[Dict]:
        """
        Search public Facebook groups for Occitanie spots

        Popular Occitanie/outdoor groups:
        - Randonnées Occitanie
        - Spots de baignade Occitanie
        - Pyrénées Randonnées
        - Lacs et Cascades du Sud
        """
        posts = []

        logger.info("🔍 Searching Facebook groups for Occitanie spots...")

        # Note: Facebook scraping requires careful handling
        # Best approaches:
        # 1. Use Facebook Graph API (requires app approval)
        # 2. Use Playwright/Puppeteer with login
        # 3. Parse public RSS feeds from pages
        # 4. Use Facebook's public search

        return posts

    def extract_location_from_post(self, post_text: str) -> Optional[str]:
        """Extract location mentions from Facebook post"""
        # Common patterns in French posts
        patterns = [
            r"(?:au|à la?|aux?)\s+([A-ZÀ-Ü][a-zà-ü\s\-']+)",
            r"(?:lac|cascade|gorges?|pic|mont)\s+(?:de\s+|d')?([A-ZÀ-Ü][a-zà-ü\s\-']+)",
            r"📍\s*([A-ZÀ-Ü][a-zà-ü\s\-']+)",
        ]

        for pattern in patterns:
            match = re.search(pattern, post_text)
            if match:
                return match.group(1).strip()

        return None

    def parse_facebook_post(self, post_element) -> Optional[Dict]:
        """Parse a Facebook post element for spot data"""
        try:
            # Extract post data
            post_text = post_element.get("text", "")
            location = self.extract_location_from_post(post_text)

            if not location:
                return None

            # Extract activities from post
            activities = []
            activity_keywords = {
                "baignade": ["baignade", "nager", "plage", "swim"],
                "randonnée": ["randonnée", "rando", "marche", "trek"],
                "escalade": ["escalade", "grimpe", "voie"],
                "camping": ["camping", "bivouac", "tente"],
            }

            post_lower = post_text.lower()
            for activity, keywords in activity_keywords.items():
                if any(kw in post_lower for kw in keywords):
                    activities.append(activity)

            return {
                "location_name": location,
                "caption": post_text[:500],  # Limit caption length
                "activities": activities,
                "source": "facebook",
                "is_real_data": True,
                "collected_at": datetime.now().isoformat(),
            }

        except Exception as e:
            logger.info(f"Error parsing post: {e}")
            return None

    def search_via_facebook_api(self) -> List[Dict]:
        """
        Use Facebook Graph API for public data
        Requires Facebook App with proper permissions
        """
        # Facebook Graph API approach
        # Requires:
        # 1. Facebook App ID
        # 2. Access Token with public_profile permission
        # 3. Page Public Content Access for public pages

        api_posts = []

        # Example API endpoints:
        # - /search?type=place&q=Occitanie
        # - /{page-id}/posts
        # - /search?type=page&q=randonnée+occitanie

        return api_posts

    def get_facebook_scraping_options(self) -> Dict:
        """Return available Facebook scraping methods"""
        return {
            "methods": [
                {
                    "name": "Facebook Graph API",
                    "description": "Official API - requires app approval",
                    "requirements": ["App ID", "Access Token", "Page permissions"],
                    "pros": ["Official", "Reliable", "No blocking"],
                    "cons": ["Approval process", "Limited to public data"],
                },
                {
                    "name": "Playwright Browser Automation",
                    "description": "Browser-based scraping with login",
                    "requirements": ["Facebook account", "Playwright setup"],
                    "pros": ["Access to more content", "Handles dynamic content"],
                    "cons": ["Slower", "Detection risk", "Login required"],
                },
                {
                    "name": "Public RSS/JSON feeds",
                    "description": "Some pages offer RSS or JSON feeds",
                    "requirements": ["None"],
                    "pros": ["No auth needed", "Fast", "Reliable"],
                    "cons": ["Very limited pages", "Basic data only"],
                },
                {
                    "name": "Facebook Embedded Posts",
                    "description": "Parse publicly embedded posts",
                    "requirements": ["Post URLs"],
                    "pros": ["Always public", "No auth"],
                    "cons": ["Need to know post URLs"],
                },
            ],
            "recommended_groups": [
                "Randonnées en Occitanie",
                "Spots de baignade Sud de la France",
                "Pyrénées - Randonnées et Refuges",
                "Lacs et Rivières Occitanie",
                "Escalade Occitanie",
                "Camping Sauvage France",
                "VTT Occitanie",
                "Canyoning Pyrénées",
            ],
            "search_queries": [
                "lac baignade occitanie",
                "cascade pyrénées orientales",
                "randonnée ariège",
                "spot secret hérault",
                "gorges tarn",
                "pic pyrénées",
            ],
        }


def main():
    """Example usage"""
    scraper = FacebookRealScraper()

    # Get available options
    options = scraper.get_facebook_scraping_options()

    logger.info("📘 Facebook Scraping Options")
    logger.info("=" * 50)

    for method in options["methods"]:
        logger.info(f"\n✅ {method['name']}")
        logger.info(f"   {method['description']}")
        logger.info(f"   Pros: {', '.join(method['pros'])}")
        logger.info(f"   Cons: {', '.join(method['cons'])}")

    logger.info(f"\n🎯 Recommended Groups to Search:")
    for group in options["recommended_groups"]:
        logger.info(f"   - {group}")

    logger.info(f"\n🔍 Suggested Search Queries:")
    for query in options["search_queries"]:
        logger.info(f"   - {query}")

    logger.info("\n💡 Next Steps:")
    logger.info("1. Choose a scraping method based on your needs")
    logger.info("2. Set up required authentication (API keys or login)")
    logger.info("3. Implement the chosen method")
    logger.info("4. Process posts through the data validation pipeline")


if __name__ == "__main__":
    main()
