requirements.txt
# Core dependencies
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlite3-to-mysql==2.1.7
sqlalchemy==2.0.25
alembic==1.13.1

# Web scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
playwright==1.41.1

# Reddit API
asyncpraw==7.7.1

# Instagram API
instagrapi==2.1.0

# Data processing
pandas==2.2.0
numpy==2.2.1
scipy==1.13.0

# Natural Language Processing
spacy==3.7.2
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl

# Geocoding
geopy==2.4.1
pyproj==3.6.1
gpxpy==1.6.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
aiofiles==0.12.0  # Compatible with asyncpraw
tenacity==8.2.3
tqdm==4.66.1

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-httpx==0.28.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
ipython==8.20.0

# Type stubs
types-requests==2.31.0.20240125
types-beautifulsoup4==4.12.0.20240106
